Move test of least_squares into tests/, and minor cleanup
